requirements.txt
#requirements.txt

streamlit
python-dotenv
azure-storage-blob
pymupdf
python-pptx
openai
langchain
langchain-openai
llama-index
tiktoken
fastapi
uvicorn
pydantic
mcp-agents
aiohttp
azure-storage-blob
langchain-community
chromadb
pydantic
asyncio